Dev tests

+++ src/endpoints/accounts-payable/accounts-payable.service.ts
@@ -1,12 +1,12 @@
 import { Injectable } from '@nestjs/common';
+import { PrismaClient } from '@prisma/client';
+import { BadRequestError } from 'src/common/errors/types/BadRequestError';
+import { ConflictError } from 'src/common/errors/types/ConflictError';
+import { NotFoundError } from 'src/common/errors/types/NotFoundError';
+import { UnauthorizedError } from 'src/common/errors/types/UnauthorizedError';
+import { Utils } from 'src/utils';
 import { CreateAccountsPayableDto } from './dto/create-accounts-payable.dto';
 import { UpdateAccountsPayableDto } from './dto/update-accounts-payable.dto';
-import { PrismaClient } from '@prisma/client';
-import { NotFoundError } from 'src/common/errors/types/NotFoundError';
-import { Utils } from 'src/utils';
-import { BadRequestError } from 'src/common/errors/types/BadRequestError';
-import { UnauthorizedError } from 'src/common/errors/types/UnauthorizedError';
-import { ConflictError } from 'src/common/errors/types/ConflictError';
 
 const prisma = new PrismaClient();
 @Injectable()
@@ -27,6 +27,10 @@ export class AccountsPayableService {
 
     if (!sheet) {
       throw new NotFoundError(`not found sheetId: ${createAccountsPayableDto.sheetId}`);
+    }
+
+    if (sheet.creatorUserId !== userId) {
+      throw new UnauthorizedError('you don\'t have permission to add this payable in this sheet')
     }
 
     const accountPayable = await prisma.accountsPayable.findFirst({
@@ -54,16 +58,13 @@ export class AccountsPayableService {
     }
   }
 
-  async findAll() {
+  async findAll(userId: number) {
     const accountsPayable = await prisma.accountsPayable.findMany({
+      where: {
+        creatorUserId: userId
+      },
       include: {
         createdBy: {
-          select: {
-            id: true,
-            name: true
-          }
-        },
-        updatedBy: {
           select: {
             id: true,
             name: true
@@ -85,7 +86,7 @@ export class AccountsPayableService {
     }))
   }
 
-  async findOne(id: number) {
+  async findOne(id: number, userId: number) {
     if (!this.utils.isNotNumber(String(id))) {
       throw new BadRequestError('invalid id')
     }
@@ -95,12 +96,6 @@ export class AccountsPayableService {
       },
       include: {
         createdBy: {
-          select: {
-            id: true,
-            name: true
-          }
-        },
-        updatedBy: {
           select: {
             id: true,
             name: true
@@ -117,6 +112,10 @@ export class AccountsPayableService {
 
     if (!accountPayable) {
       throw new NotFoundError(`not found accountPayableId: ${id}`);
+    }
+
+    if (accountPayable.creatorUserId !== userId) {
+      throw new UnauthorizedError('you don\'t have permission to access this account payable')
     }
 
     return {
@@ -137,6 +136,23 @@ export class AccountsPayableService {
       throw new UnauthorizedError(`user token invalid`);
     }
 
+    if (updateAccountsPayableDto.sheetId) {
+      const sheet = await prisma.sheets.findUnique({
+        where: {
+          id: updateAccountsPayableDto.sheetId
+        }
+      })
+
+      if (!sheet) {
+        throw new BadRequestError('sheetId invalid')
+      }
+
+      if (sheet.creatorUserId !== Number(userIdUpdate)) {
+        throw new UnauthorizedError('you don\'t have permission to add this payable in this sheet')
+      }
+      return sheet
+    }
+
     const accountPayable = await prisma.accountsPayable.findUnique({
       where: {
         id
@@ -147,24 +163,15 @@ export class AccountsPayableService {
       throw new NotFoundError(`not found accountPayableId: ${id}`);
     }
 
-    const sheet = await prisma.sheets.findUnique({
-      where: {
-        id: updateAccountsPayableDto.sheetId
-      }
-    })
-
-    if (!sheet) {
-      throw new NotFoundError(`not found sheetId: ${id}`);
+    if (accountPayable.creatorUserId !== Number(userIdUpdate)) {
+      throw new UnauthorizedError('you don\'t have permission to modify this account payable')
     }
 
     const accountPayableUpdated = await prisma.accountsPayable.update({
       where: {
         id
       },
-      data: {
-        ...updateAccountsPayableDto,
-        updaterUserId: Number(userIdUpdate)
-      }
+      data: updateAccountsPayableDto
     });
 
     return {
@@ -174,7 +181,7 @@ export class AccountsPayableService {
     }
   }
 
-  async remove(id: number) {
+  async remove(id: number, userId: number) {
     if (!this.utils.isNotNumber(String(id))) {
       throw new BadRequestError('invalid id')
     }
@@ -186,6 +193,10 @@ export class AccountsPayableService {
 
     if (!accountPayable) {
       throw new NotFoundError(`not found accountPayableId: ${id}`);
+    }
+
+    if (accountPayable.creatorUserId !== userId) {
+      throw new UnauthorizedError('you don\'t have permission to delete this account payable')
     }
 
     const accountPayableDeleted = await prisma.accountsPayable.delete({
